+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.3.0] (2026-09-15)
 
 - **Added:** Support for the Client List window (via /timers).
 - **Added:** A /scrollabletabs chat command to open the configuration window.
@@ -17,7 +17,8 @@
 
 First release. 🥳
 
-[unreleased]: https://github.com/Haselnussbomber/ScrollableTabs/compare/v1.2.0...main
+[unreleased]: https://github.com/Haselnussbomber/ScrollableTabs/compare/v1.3.0...main
+[1.3.0]: https://github.com/Haselnussbomber/ScrollableTabs/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Haselnussbomber/ScrollableTabs/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Haselnussbomber/ScrollableTabs/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Haselnussbomber/ScrollableTabs/commit/6fbf4da3c254586cb8e23329938edfc9823c4e8f